Superswaps (#18)
* include chain id and name in token and pools

* superswap prep

* add uni chain

* migrate swap to the new contract

* add uni chain rpc url settings

* superswaps wip

* superswaps flow WIP

* test build_super_swap_data

* move things around, clean up API for super wasps 🐝

* superwasp stuff

* superswap dynamic fees WIP

* clean up

* clean up bridge token usage for superswaps

* dynamic fees WIP

* clean up

* update config, latest contracts

* use pool type props insteaf of fields

* adjust pack_path for v2 swaps

* clean up route_for_swap

* clean up

* adjust bridged amout in superswap quote

* adjustments to build superswap data function

* swap adjustments

* superswap tweaks

* add get_token, export all chains by default, superswap API design

* add lisk, superwasp op -> lisk

* move things around, add test for superswap

* config lisk in CI

* clean up

* update list of env vars

* check if chains support superswap before attempting it

* clean up get_domain jazz

* clean up

* sync version of superswap

* clean up

* fixes

* ping

* move tests for swaps into a separate nb

* try fewer tests

* stabilize tests

* clean up

* try bigger slippage in tests

* implement bridging in superswaps

* test bridging

* update base settings, add note on release chores

* migrate to etherscan v2

* Superswap refactor (#20)

* update base settings, add note on release chores

* migrate to etherscan v2

* move things around

* helper fixes, wei all the way

* assorted fixes

* clean up

* update contracts for base

* update config section in readme

* clean up calc_bridged_amount

* tweak superswap quote

* clean up method name for bridge token, update docs

* tweak param name for superswap quotes

* handle quote not found for superswaps, update docs

* bring back slipstream and nfpm

* rename bridge_transfer_from_abi

* rename CONTRACT_BALANCE_FOR_V3_SWAPS to CONTRACT_BALANCE

* fetch fee right before the tx

* link to the universal router deployments in the docs

* change sig for V2_SWAP_EXACT_OUT

* Refactor amounts params (#22) re: #21

* refactor amounts params

* rename conversion helpers

* fix method name

* rename friendly_amount to as_float

* add diff + Claude report

* clean up

* clean up, add breaking changes doc

* fixes round 2

* add comment re unused destination token

* 🍯 Honey (#23)

* wip

* logging

* create wallet, check balances

* check_balances_all_chains func

* load on eth WIP, wallet WIP + doodling on honey

* add velo to the wallet

* wip

* wip

* make honey.yaml mandatory

* clean up, use sugar with honey

* integrate new test setup WIP

* clean up wallet loading

* lowercase honey

* WIP

* clean up

* update ci

* run supersim in non background more for now

* missing deps

* sleep a bit after ousdt swap

* use per token holders

* drop balance tests

* try to speed things up

* check_chain_balances in parallel

* add helpers for simnets

* clean up superswap tests

* move things around

* add honey.md

* remove wallet helper

* clean up

* add note re:private keys

* clean up

* address pending nits, add more tests

* split tests

* unsplit tests

* clean up

* clean up

* update readme

* update readme

* update README.md

+++ sugar/deposit.py
@@ -14,4 +14,4 @@ class Deposit:
     # pool to deposit into
     pool: LiquidityPool
     # amount of token0 for deposit
-    amount_token0: float
+    amount_token0: int
